requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
bcrypt==4.1.3
matplotlib==3.9.0
python-dotenv==1.0.1
supabase-client==0.4.6
